+++ libs/components/router/projects/router/src/modules/href/href.module.ts
@@ -1,13 +1,11 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { SkyAppWindowRef } from '@skyux/core';
 
 import { SkyHrefDirective } from './href.directive';
 
 @NgModule({
   declarations: [SkyHrefDirective],
   exports: [SkyHrefDirective],
-  imports: [CommonModule],
-  providers: [SkyAppWindowRef]
+  imports: [CommonModule]
 })
 export class SkyHrefModule {}