+++ libs/components/router/projects/router/src/modules/link/link.module.ts
@@ -3,10 +3,6 @@ import {
 } from '@angular/core';
 
 import {
-  SkyAppWindowRef
-} from '@skyux/core';
-
-import {
   SkyAppLinkDirective
 } from './link.directive';
 
@@ -22,9 +18,6 @@ import {
   exports: [
     SkyAppLinkDirective,
     SkyAppLinkExternalDirective
-  ],
-  providers: [
-    SkyAppWindowRef
   ]
 })
 /* istanbul ignore next */