+++ libs/components/router/projects/router/src/public-api.ts
@@ -1,0 +1,13 @@
+export * from './modules/link/link-query-params';
+export * from './modules/link/link.module';
+export * from './modules/href/href.module';
+export * from './modules/href/href-resolver';
+export * from './modules/href/href-resolver.service';
+export * from './modules/href/types/href';
+export * from './modules/href/types/href-resolver.args';
+
+// Components and directives must be exported to support Angular's "partial" Ivy compiler.
+// Obscure names are used to indicate types are not part of public API.
+export { SkyHrefDirective as λ1 } from './modules/href/href.directive';
+export { SkyAppLinkExternalDirective as λ2 } from './modules/link/link-external.directive';
+export { SkyAppLinkDirective as λ3 } from './modules/link/link.directive';